[ai] add hamburger menu with Import/Export; Export json -> share
codex prompt:

how do we go about adding support for file import/export? conceptually it seems we should start with adding an export
function. if we copy FreeOTP+, it adds a hamburger menu to the top right of the app. if there's a standard motif for ios we
can use that too. in the menu, there are options to Import and Export. use a dummy entry for Import whose handler does
nothing. For Export, it needs to offer to save a file (dunno how this works, but if there's no native file storage like in
android, we can do a share intent that allows us to pass the saved file to email, gdrive, airdrop or whatever).

so to complete this process, we need to produce a json file. it has 2 keys, the schema is roughly as follows, in yaml

  type: object
  properties:
    tokenOrder:
      type: array
      items:
        type: string
      examples:
      - ["ngrok:foo@bar.com", "github:myuser", "heroku:some@one.there"]
    tokens:
      type: array
      # note that len(tokenOrder) == len(tokens)
      items:
        type: object
        properties:
          algo:
            type: string
            examples: ["SHA1"]
          counter:
            type: integer
          digits:
            type: integer
          issuerExt:
            type: string
          label:
            type: string
          period:
            type: number
          secret:
            type: array
            items:
              type: integer
          type:
            type: string

since FreeOTP+ is derived from FreeOTP, i expect many of these properties to be shared. so we shouldn't have too hard a time
getting there, good?