[ai] add Import from json functionality
codex prompt:
ok the export seems to work. now add the importer. it should read from using the Files provider. to validate this process, we
should be able to export a json through the hamburger (i have verified this works), then read in the same json file back.

for reading, we should parse the json into the same structure that we've exported it from. should also be straightforward

---

build fails with this error

freeotp-ios/FreeOTP/TokensViewController.swift:470:33 Redundant conformance of
'TokensViewController' to protocol 'UIDocumentPickerDelegate'